+++ test/interable_extensions_test.dart
@@ -1,5 +1,6 @@
 import 'package:test/test.dart';
 import 'package:list_ext/list_ext.dart';
+import 'package:tuple/tuple.dart';
 
 void main() {
   group('Iterable', () {
@@ -103,6 +104,7 @@ void main() {
     });
 
     group('math', () {
+      // Sum
       test('sum of int elements property', () {
         final list = [_E(intVal: 1), _E(intVal: 3)];
 
@@ -114,6 +116,49 @@ void main() {
 
         expect(list.sumOfDouble((e) => e.doubleVal), 4.6);
       });
+
+      // Average
+      test('ceil avg of int elements property', () {
+        final list = [_E(intVal: 1), _E(intVal: 3)];
+
+        expect(list.avgOf((e) => e.intVal), 2);
+      });
+
+      test('float avg of int elements property', () {
+        final list = [_E(intVal: 1), _E(intVal: 4)];
+
+        expect(list.avgOf((e) => e.intVal), 2.5);
+      });
+
+      test('avg of int with negative', () {
+        final list = [_E(intVal: 1), _E(intVal: -3)];
+
+        expect(list.avgOf((e) => e.intVal), -1);
+      });
+
+      test('avg of empty', () {
+        final list = [];
+
+        expect(list.avgOf((e) => e.intVal), 0);
+      });
+
+      test('avg of double elements property', () {
+        final list = [_E(doubleVal: 1.5), _E(doubleVal: 3.2)];
+
+        expect(list.avgOfDouble((e) => e.doubleVal), 2.35);
+      });
+
+      test('avg of double with negative', () {
+        final list = [_E(doubleVal: 2.1), _E(doubleVal: -1.1)];
+
+        expect(list.avgOfDouble((e) => e.doubleVal), .5);
+      });
+
+      test('avg of double empty', () {
+        final list = [];
+
+        expect(list.avgOfDouble((e) => e.doubleVal), 0);
+      });
     });
   });
 
@@ -123,6 +168,23 @@ void main() {
 
       expect(list.sum(), 8);
     });
+
+    group('avg()', () {
+      final data = <Tuple3<String, Iterable<int>, double>>[
+        Tuple3('simple', [1, 3, 5], 3),
+        Tuple3('negative', [3, -2, 5], 2),
+        Tuple3('float', [5, 4], 4.5),
+        Tuple3('empty', [], 0),
+      ];
+
+      for (final d in data) {
+        test('average elements: ${d.item1}', () {
+          final list = d.item2;
+
+          expect(list.avg(), d.item3);
+        });
+      }
+    });
   });
 
   group('Iterable of double', () {
@@ -130,6 +192,22 @@ void main() {
       final list = [1.5, 2.3, 5.7];
 
       expect(list.sum(), 9.5);
+    });
+
+    group('avg()', () {
+      final data = <Tuple3<String, Iterable<double>, double>>[
+        Tuple3('simple', [1.6, 2.3, 5.7], 3.2),
+        Tuple3('negative', [1.3, 2.3, -5.7], -.7),
+        Tuple3('empty', [], 0),
+      ];
+
+      for (final d in data) {
+        test('average elements: ${d.item1}', () {
+          final list = d.item2;
+
+          expect(list.avg(), closeTo(d.item3, .00001));
+        });
+      }
     });
   });
 }